Remove cache from version control and create caching module #27
