+++ tests/test_html.py
@@ -1,45 +1,12 @@
-import json
-import os
-
 import ee
 
 from eerepr.html import convert_to_html
-
-
-def load_info(obj):
-    """Load client-side info for an Earth Engine object.
-
-    Info is retrieved (if available) from a local JSON file using the serialized
-    object as the key. If the data does not exist locally, it is loaded from Earth
-    Engine servers and stored for future use.
-    """
-    serialized = obj.serialize()
-
-    if not os.path.isdir("./tests/data"):
-        os.mkdir("./tests/data")
-
-    try:
-        with open("./tests/data/data.json") as src:
-            existing_data = json.load(src)
-
-    # File is missing or unreadable
-    except (FileNotFoundError, json.JSONDecodeError):
-        existing_data = {}
-        with open("./tests/data/data.json", "w") as dst:
-            json.dump(existing_data, dst)
-
-    # File exists, but info does not
-    if serialized not in existing_data:
-        with open("./tests/data/data.json", "w") as dst:
-            existing_data[serialized] = obj.getInfo()
-            json.dump(existing_data, dst)
-
-    return existing_data[serialized]
+from tests.cache import get_info
 
 
 def test_image():
     obj = ee.Image.constant(0).set("system:id", "foo")
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
     assert "Image foo (1 band)" in rep
     assert "bands: List (1 element)" in rep
@@ -53,7 +20,7 @@ def test_imagecollection():
             ee.Image.constant(1),
         ]
     ).set("test_prop", 42)
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "ImageCollection (2 elements)" in rep
@@ -63,7 +30,7 @@ def test_imagecollection():
 
 def test_feature():
     obj = ee.Feature(geom=ee.Geometry.Point([0, 0]), opt_properties={"foo": "bar"})
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Feature (Point, 1 property)" in rep
@@ -71,7 +38,7 @@ def test_feature():
 
 def test_empty_feature():
     obj = ee.Feature(None)
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Feature (0 properties)" in rep
@@ -80,7 +47,7 @@ def test_empty_feature():
 def test_featurecollection():
     feat = ee.Feature(geom=ee.Geometry.Point([0, 0]), opt_properties={"foo": "bar"})
     obj = ee.FeatureCollection([feat])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "FeatureCollection (1 element, 2 columns)" in rep
@@ -88,7 +55,7 @@ def test_featurecollection():
 
 def test_date():
     obj = ee.Date("2021-03-27T14:01:07")
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Date" in rep
@@ -97,7 +64,7 @@ def test_date():
 
 def test_filter():
     obj = ee.Filter.eq("foo", "bar")
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Filter.eq" in rep
@@ -105,7 +72,7 @@ def test_filter():
 
 def test_dict():
     obj = ee.Dictionary({"foo": "bar"})
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Object (1 property)" in rep
@@ -113,19 +80,19 @@ def test_dict():
 
 def test_list():
     short_obj = ee.List([1, 2, 3, 4])
-    short_info = load_info(short_obj)
+    short_info = get_info(short_obj)
     short_rep = convert_to_html(short_info)
     assert "[1, 2, 3, 4]" in short_rep
 
     long_obj = ee.List.sequence(0, 20, 1)
-    long_info = load_info(long_obj)
+    long_info = get_info(long_obj)
     long_rep = convert_to_html(long_info)
     assert "List (21 elements)" in long_rep
 
 
 def test_string():
     obj = ee.String("13th Warrior is an underrated movie")
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "13th Warrior is an underrated movie" in rep
@@ -133,7 +100,7 @@ def test_string():
 
 def test_number():
     obj = ee.Number(42)
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "42" in rep
@@ -141,7 +108,7 @@ def test_number():
 
 def test_point():
     obj = ee.Geometry.Point([1.112312, 2])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Point (1.11, 2.00)" in rep
@@ -149,7 +116,7 @@ def test_point():
 
 def test_multipoint():
     obj = ee.Geometry.MultiPoint([[1, 1], [2, 2]])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "MultiPoint (2 vertices)" in rep
@@ -157,7 +124,7 @@ def test_multipoint():
 
 def test_linestring():
     obj = ee.Geometry.LineString([[1, 1], [2, 2], [3, 3]])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "LineString (3 vertices)" in rep
@@ -165,7 +132,7 @@ def test_linestring():
 
 def test_multilinestring():
     obj = ee.Geometry.MultiLineString([[[0, 0], [1, 1]]])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "MultiLineString" in rep
@@ -173,7 +140,7 @@ def test_multilinestring():
 
 def test_polygon():
     obj = ee.Geometry.Polygon([[0, 0], [1, 1], [2, 2], [0, 0]])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Polygon (4 vertices)" in rep
@@ -186,7 +153,7 @@ def test_multipolygon():
             [[4, 6], [3, 2], [1, 2], [4, 6]],
         ]
     )
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "MultiPolygon (8 vertices)" in rep
@@ -194,7 +161,7 @@ def test_multipolygon():
 
 def test_linearring():
     obj = ee.Geometry.LinearRing([[0, 0], [1, 1], [2, 2], [0, 0]])
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "LinearRing (4 vertices)" in rep
@@ -202,7 +169,7 @@ def test_linearring():
 
 def test_daterange():
     obj = ee.DateRange("2020-01-01T21:01:10", "2022-03-01T14:32:11")
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "DateRange [2020-01-01 21:01:10, 2022-03-01 14:32:11]" in rep
@@ -210,7 +177,7 @@ def test_daterange():
 
 def test_typed_obj():
     obj = ee.Dictionary({"type": "Foo", "id": "bar"})
-    info = load_info(obj)
+    info = get_info(obj)
     rep = convert_to_html(info)
 
     assert "Foo bar" in rep
@@ -230,7 +197,7 @@ def test_band():
             "crs": crs,
         }
     )
-    band1_info = load_info(band1)
+    band1_info = get_info(band1)
     band1_rep = convert_to_html(band1_info)
     assert '"B1", unsigned int16, EPSG:32610, 1830x1830 px' in band1_rep
 
@@ -240,21 +207,21 @@ def test_band():
             "data_type": data_type,
         }
     )
-    band2_info = load_info(band2)
+    band2_info = get_info(band2)
     band2_rep = convert_to_html(band2_info)
     assert '"B1", unsigned int16' in band2_rep
 
 
 def test_pixel_types():
-    assert "float" in convert_to_html(load_info(ee.PixelType.float()))
-    assert "double" in convert_to_html(load_info(ee.PixelType.double()))
-    assert "signed int8" in convert_to_html(load_info(ee.PixelType.int8()))
-    assert "unsigned int8" in convert_to_html(load_info(ee.PixelType.uint8()))
-    assert "signed int16" in convert_to_html(load_info(ee.PixelType.int16()))
-    assert "unsigned int16" in convert_to_html(load_info(ee.PixelType.uint16()))
-    assert "signed int32" in convert_to_html(load_info(ee.PixelType.int32()))
-    assert "unsigned int32" in convert_to_html(load_info(ee.PixelType.uint32()))
-    assert "signed int64" in convert_to_html(load_info(ee.PixelType.int64()))
+    assert "float" in convert_to_html(get_info(ee.PixelType.float()))
+    assert "double" in convert_to_html(get_info(ee.PixelType.double()))
+    assert "signed int8" in convert_to_html(get_info(ee.PixelType.int8()))
+    assert "unsigned int8" in convert_to_html(get_info(ee.PixelType.uint8()))
+    assert "signed int16" in convert_to_html(get_info(ee.PixelType.int16()))
+    assert "unsigned int16" in convert_to_html(get_info(ee.PixelType.uint16()))
+    assert "signed int32" in convert_to_html(get_info(ee.PixelType.int32()))
+    assert "unsigned int32" in convert_to_html(get_info(ee.PixelType.uint32()))
+    assert "signed int64" in convert_to_html(get_info(ee.PixelType.int64()))
 
     custom_type = dict(type="PixelType", min=10, max=255, precision="int")
-    assert "int ∈ [10, 255]" in convert_to_html(load_info(ee.Dictionary(custom_type)))
+    assert "int ∈ [10, 255]" in convert_to_html(get_info(ee.Dictionary(custom_type)))
